backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pip install langchain-core==0.1.52 langchain-huggingface==0.0.6

qdrant-client>=1.9.0
langchain==0.1.0
langchain-groq==0.0.1
langchain-qdrant==0.1.0
langchain-huggingface==0.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.0